migrate CLI argument handling back to main.py
This is a bit of churn.  We spun absolutely everything out of main.py
in a series of steps, but it isn't required that the file have zero
responsibilities.

Here mycli/cli_args.py is removed, and its responsibilities brought
back to main.py.

Some values are also moved to constants.py to avoid circular imports.
